Simplifying text drawers
