+++ writeopia/src/commonMain/kotlin/io/writeopia/sdk/drawer/content/MessageDrawer.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
@@ -31,9 +30,8 @@ import io.writeopia.sdk.utils.ui.defaultTextStyle
  * This class contains the logic of the basic message of the SDK. As many other drawers need some
  * text in it this Drawer can be used instead of duplicating this text logic.
  *
- * This is the intended version to be used with desktop and webapp, instead o the mobile version.
  */
-class DesktopMessageDrawer(
+class MessageDrawer(
     private val modifier: Modifier = Modifier,
     private val onKeyEvent: (KeyEvent, TextFieldValue, StoryStep, Int) -> Boolean = { _, _, _, _ -> false },
     private val textStyle: @Composable (StoryStep) -> TextStyle = { defaultTextStyle(it) },
@@ -58,49 +56,42 @@ class DesktopMessageDrawer(
             }
         }
 
-        if (drawInfo.editable) {
-            BasicTextField(
-                modifier = modifier
-                    .padding(start = 16.dp)
-                    .let { modifierLet ->
-                        if (focusRequester != null) {
-                            modifierLet.focusRequester(focusRequester)
-                        } else {
-                            modifierLet
-                        }
-                    }
-                    .onKeyEvent { keyEvent ->
-                        onKeyEvent(keyEvent, inputText, step, drawInfo.position)
-                    }
-                    .onFocusChanged(onFocusChanged),
-                value = inputText,
-                onValueChange = { value ->
-                    val text = value.text
-                    val newStep = step.copy(text = text)
-
-                    inputText = if (text.contains("\n") && !allowLineBreaks) {
-                        onLineBreak(Action.LineBreak(newStep, drawInfo.position))
-
-                        val newText = text.split("\n", limit = 2)[0]
-                        TextFieldValue(newText, TextRange(newText.length))
+        BasicTextField(
+            modifier = modifier
+                .padding(start = 16.dp)
+                .let { modifierLet ->
+                    if (focusRequester != null) {
+                        modifierLet.focusRequester(focusRequester)
                     } else {
-                        onTextEdit(Action.StoryStateChange(newStep, drawInfo.position))
-                        commandHandler.handleCommand(text, newStep, drawInfo.position)
-
-                        value
+                        modifierLet
                     }
-                },
-                keyboardOptions = KeyboardOptions(
-                    capitalization = KeyboardCapitalization.Sentences
-                ),
-                textStyle = textStyle(step),
-                cursorBrush = SolidColor(MaterialTheme.colorScheme.primary)
-            )
-        } else {
-            Text(
-                text = step.text ?: "",
-                modifier = modifier.padding(vertical = 5.dp),
-            )
-        }
+                }
+                .onKeyEvent { keyEvent ->
+                    onKeyEvent(keyEvent, inputText, step, drawInfo.position)
+                }
+                .onFocusChanged(onFocusChanged),
+            value = inputText,
+            onValueChange = { value ->
+                val text = value.text
+                val newStep = step.copy(text = text)
+
+                inputText = if (text.contains("\n") && !allowLineBreaks) {
+                    onLineBreak(Action.LineBreak(newStep, drawInfo.position))
+
+                    val newText = text.split("\n", limit = 2)[0]
+                    TextFieldValue(newText, TextRange(newText.length))
+                } else {
+                    onTextEdit(Action.StoryStateChange(newStep, drawInfo.position))
+                    commandHandler.handleCommand(text, newStep, drawInfo.position)
+
+                    value
+                }
+            },
+            keyboardOptions = KeyboardOptions(
+                capitalization = KeyboardCapitalization.Sentences
+            ),
+            textStyle = textStyle(step),
+            cursorBrush = SolidColor(MaterialTheme.colorScheme.primary)
+        )
     }
 }
