+++ writeopia/src/jvmMain/kotlin/io/writeopia/sdk/previews/drawer/content/CheckItemDrawerStepPreview.kt
@@ -2,7 +2,7 @@ package io.writeopia.sdk.previews.drawer.content
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
-import io.writeopia.sdk.drawer.content.DesktopMessageDrawer
+import io.writeopia.sdk.drawer.content.MessageDrawer
 import io.writeopia.sdk.drawer.content.checkItemDrawer
 import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.models.story.StoryStep
@@ -11,7 +11,7 @@ import io.writeopia.sdk.models.story.StoryTypes
 @Preview
 @Composable
 fun CheckItemDrawerStepPreview() {
-    checkItemDrawer(messageDrawer = { DesktopMessageDrawer() }).Step(
+    checkItemDrawer(messageDrawer = { MessageDrawer() }).Step(
         step = StoryStep(
             type = StoryTypes.CHECK_ITEM.type,
             text = "This is a check item"
