+++ writeopia/src/jvmMain/kotlin/io/writeopia/sdk/previews/drawer/content/HeaderDrawerStepPreview.kt
@@ -10,8 +10,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import io.writeopia.sdk.drawer.content.MessageDrawer
 import io.writeopia.sdk.drawer.content.HeaderDrawer
-import io.writeopia.sdk.drawer.content.MobileMessageDrawer
 import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.models.story.Decoration
 import io.writeopia.sdk.models.story.StoryStep
@@ -30,7 +30,7 @@ fun HeaderDrawerStepPreview() {
 
     HeaderDrawer(
         drawer = {
-            MobileMessageDrawer(
+            MessageDrawer(
                 modifier = Modifier.align(Alignment.BottomStart).padding(start = 16.dp, bottom = 16.dp),
                 onTextEdit = { },
                 onLineBreak = { },
@@ -57,7 +57,7 @@ fun HeaderDrawerStepPreviewNoColor() {
 
     HeaderDrawer(
         drawer = {
-            MobileMessageDrawer(
+            MessageDrawer(
                 modifier = Modifier.align(Alignment.BottomStart).padding(start = 16.dp, bottom = 16.dp),
                 onTextEdit = { },
                 onLineBreak = { },
