+++ writeopia/src/jvmMain/kotlin/io/writeopia/sdk/previews/drawer/content/SimpleMessageDrawerPreview.kt
@@ -3,7 +3,7 @@ package io.writeopia.sdk.previews.drawer.content
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.focus.FocusRequester
-import io.writeopia.sdk.drawer.content.DesktopMessageDrawer
+import io.writeopia.sdk.drawer.content.MessageDrawer
 import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.models.story.StoryStep
 import io.writeopia.sdk.models.story.StoryTypes
@@ -11,7 +11,7 @@ import io.writeopia.sdk.models.story.StoryTypes
 @Preview
 @Composable
 fun DesktopMessageDrawerPreview() {
-    DesktopMessageDrawer(focusRequester = FocusRequester()).Step(
+    MessageDrawer(focusRequester = FocusRequester()).Step(
         step = StoryStep(text = "Some text", type = StoryTypes.TEXT.type),
         drawInfo = DrawInfo()
     )
