+++ writeopia/src/jvmMain/kotlin/io/writeopia/sdk/previews/drawer/content/SwipeMessageDrawer.kt
@@ -3,7 +3,7 @@ package io.writeopia.sdk.previews.drawer.content
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.focus.FocusRequester
-import io.writeopia.sdk.drawer.content.DesktopMessageDrawer
+import io.writeopia.sdk.drawer.content.MessageDrawer
 import io.writeopia.sdk.drawer.content.swipeMessageDrawer
 import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.models.story.StoryTypes
@@ -16,7 +16,7 @@ private fun SwipeMessageDrawerPreview() {
 
     swipeMessageDrawer(
         messageDrawer = {
-            DesktopMessageDrawer(focusRequester = focusRequester)
+            MessageDrawer(focusRequester = focusRequester)
         },
         focusRequester = focusRequester
     ).Step(
