+++ writeopia/src/jvmMain/kotlin/io/writeopia/sdk/previews/drawer/content/UnOrderedListItemDrawer.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import io.writeopia.sdk.drawer.content.DesktopMessageDrawer
+import io.writeopia.sdk.drawer.content.MessageDrawer
 import io.writeopia.sdk.drawer.content.unOrderedListItemDrawer
 import io.writeopia.sdk.model.draw.DrawInfo
 import io.writeopia.sdk.models.story.StoryStep
@@ -29,7 +29,7 @@ private fun UnOrderedListItemPreview() {
         modifier,
         focusRequester = focusRequester,
         messageDrawer = {
-            DesktopMessageDrawer(focusRequester = focusRequester)
+            MessageDrawer(focusRequester = focusRequester)
         }
     ).Step(
         StoryStep(type = StoryTypes.UNORDERED_LIST_ITEM.type, text = "Item1"),
